Move FileSystem to root
