+++ src/fairseq2/recipes/asr/_eval.py
@@ -16,6 +16,7 @@ from typing_extensions import override
 from fairseq2.context import RuntimeContext
 from fairseq2.datasets import LengthBatching, SyncMode
 from fairseq2.datasets.asr import GENERIC_ASR_DATASET_FAMILY, AsrDataset, AsrReadOptions
+from fairseq2.file_system import FileMode
 from fairseq2.gang import Gangs
 from fairseq2.models.asr import AsrModel
 from fairseq2.models.seq2seq import Seq2SeqBatch
@@ -38,7 +39,6 @@ from fairseq2.recipes.config import (
     TextTokenizerSection,
 )
 from fairseq2.typing import CPU
-from fairseq2.utils.file import FileMode
 from fairseq2.utils.rng import manual_seed
 from fairseq2.utils.structured import structure
 from fairseq2.utils.validation import validate
