+++ src/fairseq2/recipes/mt/_eval.py
@@ -22,6 +22,7 @@ from fairseq2.datasets.parallel_text import (
     ParallelTextDataset,
     ParallelTextReadOptions,
 )
+from fairseq2.file_system import FileMode
 from fairseq2.gang import Gangs
 from fairseq2.generation import BeamSearchConfig, Seq2SeqGenerator
 from fairseq2.generation.text import SequenceToTextConverter
@@ -57,7 +58,6 @@ from fairseq2.recipes.config import (
     TextTokenizerSection,
 )
 from fairseq2.typing import CPU
-from fairseq2.utils.file import FileMode
 from fairseq2.utils.rng import manual_seed
 from fairseq2.utils.structured import structure
 from fairseq2.utils.validation import validate
